+++ src/objects.js
@@ -1870,6 +1870,7 @@ SpriteMorph.prototype.init = function (globals) {
 
     this.primitivesCache = {}; // not to be serialized (!)
     this.paletteCache = {}; // not to be serialized (!)
+    this.emptyCategoriesCache = {}; // not to be serialized;
     this.rotationOffset = ZERO; // not to be serialized (!)
     this.idx = 0; // not to be serialized (!) - used for de-serialization
 
@@ -1935,6 +1936,7 @@ SpriteMorph.prototype.fullCopy = function (forClone) {
     c.freqPlayer = null;
     c.primitivesCache = {};
     c.paletteCache = {};
+    c.emptyCategoriesCache = {};
     c.imageData = {};
     c.cachedColorDimensions = c.color[this.penColorModel]();
     arr = [];
@@ -3011,6 +3013,14 @@ SpriteMorph.prototype.freshPalette = function (category) {
         searchButton,
         makeButton;
 
+    function isEmptyCategory(blocks) {
+        return blocks.filter(each =>
+            each instanceof BlockMorph && !myself.isHidingBlock(each)
+        ).length === 0;
+    }
+
+    this.emptyCategoriesCache = {};
+
     palette.owner = this;
     palette.padding = unit / 2;
     palette.color = this.paletteColor;
@@ -3088,7 +3098,7 @@ SpriteMorph.prototype.freshPalette = function (category) {
         // In a Unified Palette custom blocks appear following each category,
         // but there is only 1 make a block button (at the end).
         showCategories = this.parentThatIsA(IDE_Morph).scene.showCategories;
-        blocks = SpriteMorph.prototype.allCategories().reduce( // +++
+        blocks = SpriteMorph.prototype.allCategories().reduce(
             (blocks, category) => {
                 let header = [this.categoryText(category), '-'],
                     primitives = this.getPrimitiveTemplates(category),
@@ -3101,6 +3111,11 @@ SpriteMorph.prototype.freshPalette = function (category) {
                 if (!showCategories && category !== 'variables') {
                     primitives = primitives.filter(each =>
                         each !== '-' && each !== '=');
+                }
+
+                // update the empty categories cache
+                if (isEmptyCategory(primitives) &&  isEmptyCategory(customs)) {
+                    this.emptyCategoriesCache[category] = true;
                 }
 
                 return blocks.concat(
@@ -7772,6 +7787,7 @@ StageMorph.prototype.init = function (globals) {
     this.keysPressed = {}; // for handling keyboard events, do not persist
     this.primitivesCache = {}; // not to be serialized (!)
     this.paletteCache = {}; // not to be serialized (!)
+    this.emptyCategoriesCache = {}; // not to be serialized (!)
     this.lastAnswer = ''; // last user input, do not persist
     this.activeSounds = []; // do not persist